requirements.txt
aiogram==3.13.1
apscheduler==3.10.4
SQLAlchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
pydantic==2.11.1
openai==1.47.0
requests==2.32.3
tzlocal==5.2
uvicorn==0.30.6
fastapi==0.114.0
Jinja2==3.1.4
